// iot-service/iot-angular/src/sdk/core/models/sensor-channel-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface SensorChannelResponseDto {
  aggregation?: string;
  alertSuppressionWindow?: number;
  createdAt: string;
  idSensor: string;
  idSensorChannel: string;
  idSensorType: string;
  maxThreshold?: number;
  metricCode: string;
  minThreshold?: number;
  multiplier?: number;

  /**
   * Value offset for calibration (renamed from offset)
   */
  offsetValue?: number;
  precision?: number;
  registerAddress?: number;

  /**
   * Sensor details
   */
  sensor?: {
};

  /**
   * Sensor type details
   */
  sensorType?: {
};
  unit?: string;
  updatedAt: string;
}